Most of Minecraft 1.3 protocol changes Still needs support for packets 0x37, 0xCB, and 0xCC

+++ include/packet/packet_block_change.h
@@ -1,2 +1,2 @@
 
-extern void packet_send_block_change(struct bedrock_client *client, int32_t x, uint8_t y, int32_t z, uint8_t id, uint8_t data);
+extern void packet_send_block_change(struct bedrock_client *client, int32_t x, uint8_t y, int32_t z, uint16_t id, uint8_t data);
